Show translated word (term) for definitions from Wikipedia
E.g.:
Taksim-Platz - Taksim Square
Petersplatz - St. Peter's Square
Karlspreis - Charlemagne Prize

In these cases it's not that important, as the thing the definition
references is fairly obvious, but in other cases it's not that
obvious. Also, if the initial sentences are TL;DR, the translated
term allows the reader to quickly understand (assuming the
translated term is known to the reader)