Handle partial lookup where selection covers two nodes
Well, when the selection ends on a node boundary, anyway

E.g. the text is 'some <a href="...">information about ...'
If the user selects 'information', Firefox treats that as a selection over 2
nodes (which is the anchor and which is the focus depends on the direction the
mouse was dragged to make the selection)

Previously a partial lookup on such a selection was silently ignored.
Now it should be handled correctly.